refactor: lift AVS consumer contract into pkg/avsclient
The types REST handlers depend on to talk to the aggregator (OperatorView,
OperatorLister, WithdrawService, WithdrawRequest/Result, ServerDeps) used
to live inside aggregator/rest/. That coupled the contract to its current
consumer — a future repo split that moved REST into a private gateway
repo would have dragged the type definitions along with it, leaving any
other consumer (CLI, MCP, second gateway) without a published contract.

Lift them verbatim to pkg/avsclient/ as the neutral home:
  - OperatorView, OperatorLister
  - WithdrawService, WithdrawRequest, WithdrawResult
  - ServerDeps -> Surface (drops the REST-specific name)

Aggregator now implements avsclient.Surface via AvsClientSurface() (was
buildRestDeps). The adapters move to aggregator/avsclient_adapter.go
(renamed from rest_adapter.go to match what they actually do). REST
imports pkg/avsclient for the types; the contract definitions are no
longer in its package.

No behavior change — every signature lift is purely a rename. The build
is green; aggregator + aggregator/rest tests pass.

What this unblocks: a future split that pulls aggregator/rest/ into a
separate (private) repo can do so by depending on this module via
normal go-modules import. The remaining coupling is
aggregator/http_server.go mounting REST directly — that bootstrap
untangle is the next surgical move when ready to commit to the split.
