+++ aggregator/avsclient_adapter.go
@@ -4,18 +4,22 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/AvaProtocol/EigenLayer-AVS/aggregator/rest"
 	"github.com/AvaProtocol/EigenLayer-AVS/model"
+	"github.com/AvaProtocol/EigenLayer-AVS/pkg/avsclient"
 	avsproto "github.com/AvaProtocol/EigenLayer-AVS/protobuf"
 	"github.com/ethereum/go-ethereum/common"
 )
 
-// buildRestDeps gathers every dependency the REST layer needs from the
-// rest of the aggregator's state. Called from startHttpServer after
-// startRpcServer + startTaskEngine have populated agg.smartWalletRpc,
+// AvsClientSurface bundles every dependency a consumer needs into an
+// avsclient.Surface. Called from startHttpServer after startRpcServer
+// + startTaskEngine have populated agg.smartWalletRpc,
 // agg.priceService, and agg.rpcServer.
-func (agg *Aggregator) buildRestDeps() rest.ServerDeps {
-	return rest.ServerDeps{
+//
+// Exported so a future out-of-process consumer (e.g. a private gateway
+// binary) can construct a Surface by importing this package, instead
+// of every consumer rewriting the adapter glue.
+func (agg *Aggregator) AvsClientSurface() avsclient.Surface {
+	return avsclient.Surface{
 		Operators:             &operatorPoolAdapter{pool: agg.operatorPool},
 		SmartWalletRpc:        agg.smartWalletRpc,
 		SmartWalletRpcByChain: agg.smartWalletRpcByChain,
@@ -25,23 +29,21 @@ func (agg *Aggregator) buildRestDeps() rest.ServerDeps {
 }
 
 // operatorPoolAdapter adapts the aggregator's *OperatorPool to the
-// rest.OperatorLister interface so the REST package doesn't need to
-// import aggregator (which would cycle). The adapter copies the fields
-// the REST layer actually surfaces; the internal *OperatorNode keeps
-// extra columns (RemoteIP, MetricsPort) that aren't part of the public
-// API shape.
+// avsclient.OperatorLister interface. The adapter copies the fields
+// consumers surface; the internal *OperatorNode keeps extra columns
+// (RemoteIP, MetricsPort) that aren't part of the public shape.
 type operatorPoolAdapter struct {
 	pool *OperatorPool
 }
 
-func (a *operatorPoolAdapter) List() []rest.OperatorView {
+func (a *operatorPoolAdapter) List() []avsclient.OperatorView {
 	if a == nil || a.pool == nil {
 		return nil
 	}
 	nodes := a.pool.GetAll()
-	out := make([]rest.OperatorView, 0, len(nodes))
+	out := make([]avsclient.OperatorView, 0, len(nodes))
 	for _, n := range nodes {
-		out = append(out, rest.OperatorView{
+		out = append(out, avsclient.OperatorView{
 			Address:         n.Address,
 			Name:            n.Name,
 			Version:         n.Version,
@@ -56,21 +58,20 @@ func (a *operatorPoolAdapter) List() []rest.OperatorView {
 }
 
 // withdrawServiceAdapter adapts RpcServer.ExecuteWithdraw to the
-// rest.WithdrawService interface. The full UserOp pipeline (bundler
-// client, paymaster, balance + gas estimation, max-amount calc) stays
-// inside the aggregator package — the REST layer just supplies the
-// authed user + request shape and translates the protobuf response
-// back to the OpenAPI WithdrawResponse.
+// avsclient.WithdrawService interface. The full UserOp pipeline
+// (bundler client, paymaster, balance + gas estimation, max-amount
+// calc) stays inside the aggregator package — consumers just supply
+// the authed user + request shape and get the hashes back.
 type withdrawServiceAdapter struct {
 	rpc *RpcServer
 }
 
-func (a *withdrawServiceAdapter) Withdraw(ctx context.Context, req rest.WithdrawRequest) (rest.WithdrawResult, error) {
+func (a *withdrawServiceAdapter) Withdraw(ctx context.Context, req avsclient.WithdrawRequest) (avsclient.WithdrawResult, error) {
 	if a == nil || a.rpc == nil {
-		return rest.WithdrawResult{}, fmt.Errorf("withdraw service not initialized")
+		return avsclient.WithdrawResult{}, fmt.Errorf("withdraw service not initialized")
 	}
 	if !common.IsHexAddress(req.Owner) {
-		return rest.WithdrawResult{}, fmt.Errorf("owner is not a valid hex address")
+		return avsclient.WithdrawResult{}, fmt.Errorf("owner is not a valid hex address")
 	}
 	user := &model.User{Address: common.HexToAddress(req.Owner)}
 	// SmartAccountAddress is intentionally not derived here — the
@@ -87,9 +88,9 @@ func (a *withdrawServiceAdapter) Withdraw(ctx context.Context, req rest.Withdraw
 
 	resp, err := a.rpc.ExecuteWithdraw(ctx, user, payload)
 	if err != nil {
-		return rest.WithdrawResult{}, err
+		return avsclient.WithdrawResult{}, err
 	}
-	return rest.WithdrawResult{
+	return avsclient.WithdrawResult{
 		UserOpHash:      resp.GetUserOpHash(),
 		TransactionHash: resp.GetTransactionHash(),
 		Status:          resp.GetStatus(),
